Revert "completed search feature"
