+++ bookshelf/src/components/BookCardDetailModal.js
@@ -3,9 +3,10 @@ import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 import Modal from "@mui/material/Modal";
-import { Container } from "@mui/material";
+
 import FormControl from "@mui/material/FormControl";
 import CardMedia from "@mui/material/CardMedia";
+
 import TextField from "@mui/material/TextField";
 import axios from "axios";
 import { UserContext } from "../UserContext";
@@ -16,7 +17,7 @@ const style = {
   left: "50%",
   transform: "translate(-50%, -50%)",
   width: 600,
-  height: 750,
+  height: 530,
   bgcolor: "background.paper",
   border: "2px solid #000",
   boxShadow: 24,
@@ -76,18 +77,7 @@ export default function BookCardDetailModal({ bookData, onBorrow }) {
                 mt: 4,
               }}
             >
-              <Box position="absolute" top="-5%" left="60%">
-                <Container>
-                  <CardMedia
-                    component="img"
-                    alt="Book Card"
-                    sx={{ width: 160 }}
-                    image={volumeInfo?.imageLinks.thumbnail}
-                  />
-                </Container>
-              </Box>
               <TextField
-                sx={{ marginTop: 30 }}
                 id="outlined-read-only-input"
                 label="Title"
                 defaultValue={volumeInfo?.title}
@@ -96,7 +86,6 @@ export default function BookCardDetailModal({ bookData, onBorrow }) {
                 }}
               />
               <TextField
-                sx={{ marginTop: 30 }}
                 id="outlined-read-only-input"
                 label="Author"
                 defaultValue={volumeInfo?.authors}
